Rehome acquire_repo with other repo functions
